adjust volume applet to use pulseaudio by default
if the user installs EVERY package as listed in chrootcfg-systemd.yaml, he also installs pulseaudio. but this break the volume widget in the panel of lxqt. see here for an exemplary post: https://forum.manjaro.org/t/manjaro-lxqt-17-0-net-edition/19023/136

this commit changes the volume applet to use pulseaudio by default. since openrc offers pulseaudio as well, we should change this!